Add some more stuff to ClickerCookieSaveData
Little issue i'm encountering right now I figured I would talk about: I
have no idea how saving is going to work. To be honest, the old upgrade
system was kinda built around making it easy to save but now it's gonna
be a massive pain in the butt to work out AND make it extensible for
possible future updates. We'll see...